add kaliop migration and docs, add composer requirements and suggestions

See merge request contextualcode/ezplatform-content-variables!1

+++ src/bundle/Controller/CollectionController.php
@@ -2,16 +2,19 @@
 
 namespace ContextualCode\EzPlatformContentVariablesBundle\Controller;
 
+use ContextualCode\EzPlatformContentVariablesBundle\Entity\Collection;
+use ContextualCode\EzPlatformContentVariablesBundle\Form\Data\ItemsSelection;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
-use ContextualCode\EzPlatformContentVariablesBundle\Entity\Collection;
-use ContextualCode\EzPlatformContentVariablesBundle\Form\Data\ItemsSelection;
 
+/**
+ * @Route("/content_variables/collection", name="content_variables.collection.")
+ */
 class CollectionController extends BaseController
 {
     /**
-     * @Route("/content_variables/collection/new", name="content_variables.collection.new")
+     * @Route("/new", name="new")
      */
     public function createAction(Request $request): Response
     {
@@ -19,7 +22,7 @@ class CollectionController extends BaseController
     }
 
     /**
-     * @Route("/content_variables/collection/edit/{id}", name="content_variables.collection.edit", defaults={"id"=null})
+     * @Route("/edit/{id}", name="edit", defaults={"id"=null})
      */
     public function editAction(Request $request, Collection $collection): Response
     {
@@ -38,11 +41,12 @@ class CollectionController extends BaseController
             'collection' => $collection,
             'form' => $form->createView(),
         ];
+
         return $this->render('@ezdesign/content_variable/collection/edit.html.twig', $params);
     }
 
     /**
-     * @Route("/content_variables/collections", name="content_variables.collection.list")
+     * @Route("/list", name="list")
      */
     public function listAction(): Response
     {
@@ -53,11 +57,12 @@ class CollectionController extends BaseController
             'collections' => $collections,
             'form' => $form->createView(),
         ];
+
         return $this->render('@ezdesign/content_variable/collection/list.html.twig', $params);
     }
 
     /**
-     * @Route("/content_variables/collection/bulk_delete", name="content_variables.collection.bulk_delete")
+     * @Route("/bulk_delete", name="bulk_delete")
      */
     public function bulkDeleteAction(Request $request): Response
     {
@@ -81,10 +86,13 @@ class CollectionController extends BaseController
             }
 
             $collection = $this->collectionHandler->find($collectionId);
-            $this->collectionHandler->delete($collection);
+            if ($collection) {
+                $this->collectionHandler->delete($collection);
+            }
 
-            $params = ['%name%' => $collection->getName()];
-            $message = $this->getTranslatedMessage('collection.delete.success', $params);
+            $message = $this->getTranslatedMessage('collection.delete.success', [
+                '%name%' => $collection ? $collection->getName() : $collectionId,
+            ]);
             $this->notificationHandler->success($message);
         }
     }
@@ -93,6 +101,7 @@ class CollectionController extends BaseController
     {
         $key = $collection->isNew() ? 'collection.new.success' : 'collection.edit.success';
         $params = ['%name%' => $collection->getName()];
+
         return $this->getTranslatedMessage($key, $params);
     }
 }
